Splitting secondary_validation (#77)
